Move notifier to another location since services isn't being ignored by sonar
