+++ VideoWeb/VideoWeb.UnitTests/Controllers/ConsultationController/InviteToConsultationTests.cs
@@ -14,7 +14,7 @@ using VideoWeb.Contract.Request;
 using VideoWeb.Controllers;
 using VideoWeb.EventHub.Hub;
 using VideoApi.Contract.Responses;
-using VideoWeb.Services;
+using VideoWeb.Helpers;
 using VideoWeb.UnitTests.Builders;
 
 namespace VideoWeb.UnitTests.Controllers.ConsultationController
