+++ VideoWeb/VideoWeb.UnitTests/Controllers/ConsultationController/RespondToConsultationTests.cs
@@ -19,9 +19,9 @@ using VideoWeb.Contract.Request;
 using VideoWeb.Contract.Responses;
 using VideoWeb.Controllers;
 using VideoWeb.EventHub.Hub;
+using VideoWeb.Helpers;
 using VideoWeb.Mappings;
 using VideoWeb.Mappings.Requests;
-using VideoWeb.Services;
 using VideoWeb.UnitTests.Builders;
 using ConsultationAnswer = VideoWeb.Common.Models.ConsultationAnswer;
