+++ VideoWeb/VideoWeb/Controllers/ConsultationsController.cs
@@ -12,7 +12,7 @@ using VideoWeb.EventHub.Models;
 using VideoWeb.Mappings;
 using VideoApi.Client;
 using VideoApi.Contract.Requests;
-using VideoWeb.Services;
+using VideoWeb.Helpers;
 using ConsultationAnswer = VideoWeb.Common.Models.ConsultationAnswer;
 
 namespace VideoWeb.Controllers
