+++ VideoWeb/VideoWeb/Extensions/ConfigureServicesExtensions.cs
@@ -34,7 +34,6 @@ using VideoWeb.Middleware;
 using BookingsApi.Client;
 using UserApi.Client;
 using VideoApi.Client;
-using VideoWeb.Services;
 using VideoWeb.Swagger;
 
 namespace VideoWeb.Extensions
